fix: Resolve cyclic import between `fs_core.ts` <-> `fs_fd.ts`
Close https://github.com/bjorn3/browser_wasi_shim/issues/58
