+++ src/index.ts
@@ -2,12 +2,14 @@ import WASI from "./wasi.js";
 export { WASI };
 
 export { Fd } from "./fd.js";
-export { File, SyncOPFSFile, Directory } from "./fs_core.js";
 export {
+  File,
+  SyncOPFSFile,
+  Directory,
   OpenFile,
   OpenDirectory,
   OpenSyncOPFSFile,
   PreopenDirectory,
-} from "./fs_fd.js";
+} from "./fs_core.js";
 export { strace } from "./strace.js";
 export * as wasi from "./wasi_defs.js";
